Add size attribute to lmv-viewer component
This change adds the ability for users to update the canvas size of lmv at
start-up and in run-time.

lmv-viewer will now track changes in attribute 'size'
size has a format of <width>x<height>. Example: size="800x600"
The idea behind this approach is to have an atomic operation for size change,
instead of separate changes to width and height which I believe will
probably come one after the other in most use cases.

In addition, the index page has been updated with options to change the
canvas size in runtime.